Stage 2 GPU OOM fix: fp16 weights + SDPA attention
Eager attention on V-JEPA-2 ViT-L's 8192-token sequence materializes
an 8192x8192 attention matrix per layer (~2 GB fp16). With weights
in fp32 (~1.2 GB) plus residual activations plus the predictor head,
peak GPU usage tipped over T4's 16 GB.

Switch to fp16 weights and attn_implementation='sdpa' (PyTorch's
memory-efficient scaled-dot-product kernel — O(N) memory instead of
O(N^2)). Also clear the cuda cache between batches and explicitly
free intermediate tensors. Should comfortably fit T4 now.

Falls back gracefully if older transformers builds reject the
attn_implementation kwarg or use the legacy `torch_dtype` name.

+++ lewm_brain/stages/stage2_features.py
@@ -41,7 +41,7 @@ def run(
     print(f"[stage2] loading {hf_id} (init={init})")
     import torch
     model, processor = features.load_vjepa2(
-        hf_id, init=init, seed=seed, dtype="float32",
+        hf_id, init=init, seed=seed, dtype="float16",
     )
     device = "cuda" if torch.cuda.is_available() else "cpu"
     print(f"[stage2] device={device}, model dtype={next(model.parameters()).dtype}")
